WorkmanPartition: track object relationship states
In addition to tracking the state of each object, we also have to track
the states of the _relationships_ between objects.

As an example, consider partitions P1 and P2, and consumer C.  Assume
that P1, P2, and C all have state WORKMAN_STATE_ALL.  In the active
state, C is associated with P1, but in the persitent state, C is instead
associated with P2.  Just storing an object state isn't enough to
represent that relationship.  We must also store the state of the
relationship itself.

This patch stores the relationship states in the WorkmanPartition class
by splitting the object properties up into "*_active" and
"*_persistent" variants.